misc cleanup

See merge request graphviz/graphviz!1504

+++ lib/sfio/sfseek.c
@@ -52,7 +52,7 @@ Sfoff_t sfseek(Sfio_t * f, Sfoff_t p, int type)
 
     /* set and initialize the stream to a definite mode */
     if ((int) SFMODE(f, local) != (mode = f->mode & SF_RDWR)) {
-	int flags = f->flags;
+	unsigned short flags = f->flags;
 
 	if (hardseek & SF_PUBLIC)	/* seek ptr must follow file descriptor */
 	    f->flags |= SF_SHARE | SF_PUBLIC;
@@ -87,7 +87,7 @@ Sfoff_t sfseek(Sfio_t * f, Sfoff_t p, int type)
     SFLOCK(f, local);
 
     /* clear error and eof bits */
-    f->flags &= ~(SF_EOF | SF_ERROR);
+    f->flags &= (unsigned short)~(SF_EOF | SF_ERROR);
 
     while (f->flags & SF_STRING) {
 	SFSTRSIZE(f);
